+++ src/commonMain/kotlin/number/NonZeroInt.kt
@@ -10,11 +10,12 @@ import kotlinx.serialization.SerializationException
 import kotools.types.Package
 import kotools.types.collection.NotEmptySet
 import kotools.types.collection.notEmptySetOf
+import kotools.types.experimental.ExperimentalKotoolsTypesApi
 import kotools.types.experimental.ExperimentalRangeApi
+import kotools.types.experimental.NotEmptyRange
 import kotools.types.internal.ExperimentalSince
 import kotools.types.internal.KotoolsTypesVersion
 import kotools.types.internal.Since
-import kotools.types.range.NotEmptyRange
 import kotools.types.text.NotBlankString
 import kotools.types.text.toNotBlankString
 
@@ -52,6 +53,7 @@ public sealed interface NonZeroInt : AnyInt {
         /** The negative range of values a [NonZeroInt] can have. */
         @ExperimentalRangeApi
         @ExperimentalSince(KotoolsTypesVersion.V4_2_0)
+        @OptIn(ExperimentalKotoolsTypesApi::class)
         public val negativeRange: NotEmptyRange<StrictlyNegativeInt> by lazy(
             StrictlyNegativeInt.Companion::range
         )
@@ -59,6 +61,7 @@ public sealed interface NonZeroInt : AnyInt {
         /** The positive range of values a [NonZeroInt] can have. */
         @ExperimentalRangeApi
         @ExperimentalSince(KotoolsTypesVersion.V4_2_0)
+        @OptIn(ExperimentalKotoolsTypesApi::class)
         public val positiveRange: NotEmptyRange<StrictlyPositiveInt> by lazy(
             StrictlyPositiveInt.Companion::range
         )