+++ src/commonTest/kotlin/experimental/NotEmptyRangeTest.kt
@@ -3,20 +3,15 @@
  * Use of this source code is governed by the MIT license.
  */
 
-package kotools.types.range
+package kotools.types.experimental
 
-import kotools.types.experimental.ExclusiveBound
-import kotools.types.experimental.ExperimentalKotoolsTypesApi
-import kotools.types.experimental.ExperimentalRangeApi
-import kotools.types.experimental.InclusiveBound
 import kotlin.random.Random
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertTrue
 
-@ExperimentalRangeApi
-@OptIn(ExperimentalKotoolsTypesApi::class)
+@ExperimentalKotoolsTypesApi
 class NotEmptyRangeTest {
     @Test
     fun notEmptyRangeOf_should_return_a_range_starting_with_the_first_bound() {