+++ src/commonTest/kotlin/number/StrictlyNegativeIntTest.kt
@@ -15,7 +15,7 @@ import kotools.types.Package
 import kotools.types.experimental.ExperimentalKotoolsTypesApi
 import kotools.types.experimental.ExperimentalRangeApi
 import kotools.types.experimental.InclusiveBound
-import kotools.types.range.NotEmptyRange
+import kotools.types.experimental.NotEmptyRange
 import kotools.types.shouldBeNotNull
 import kotools.types.shouldEqual
 import kotools.types.shouldFailWithIllegalArgumentException