+++ src/commonTest/kotlin/number/StrictlyPositiveIntTest.kt
@@ -15,8 +15,8 @@ import kotools.types.Package
 import kotools.types.experimental.ExperimentalKotoolsTypesApi
 import kotools.types.experimental.ExperimentalRangeApi
 import kotools.types.experimental.InclusiveBound
+import kotools.types.experimental.NotEmptyRange
 import kotools.types.internal.unexpectedCreationFailure
-import kotools.types.range.NotEmptyRange
 import kotools.types.shouldBeNotNull
 import kotools.types.shouldEqual
 import kotools.types.shouldFailWithIllegalArgumentException